[Silver III] Title: 주유소, Time: 356 ms, Memory: 30416 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -20,7 +20,7 @@ public class Main {
 		}
 		
 		int lowPrice = price[0];
-		int sum = distance[0] * lowPrice;
+		long sum = distance[0] * lowPrice;
 		
 		for (int i = 1; i < n - 1; i++) {
 			lowPrice = Math.min(lowPrice, price[i]);
